refactor(GearChart): 1475行→1028行に分割
- SegmentedControl → components/ui/SegmentedControl.tsx
- ChartCenterDisplay → components/charts/ChartCenterDisplay.tsx
- ChartSummaryFooter/SummaryStatCard → components/charts/ChartSummaryFooter.tsx
- CHART_CONFIG/FONT_SIZES/UL_BADGE_COLORS/formatValue/getItemValue → utils/chartConfig.ts
